+++ src/tbm_utils/decorators.py
@@ -1,8 +1,13 @@
 __all__ = [
-	'cast_to_list'
+	'cast_to_list',
+	'datareader',
 ]
 
+import functools
+
 import wrapt
+
+from .io import DataReader
 
 
 def cast_to_list(position):
@@ -18,3 +23,21 @@ def cast_to_list(position):
 		return wrapped(*args, **kwargs)
 
 	return wrapper
+
+
+def datareader(wrapped=None, *, position=0):
+	"""Cast the positional argument at given position to :class:`DataReader`."""
+
+	if wrapped is None:  # pragma: nocover
+		return functools.partial(datareader, position=position)
+
+	@wrapt.decorator
+	def wrapper(wrapped, instance, args, kwargs):
+		if not isinstance(args[position], DataReader):
+			args = list(args)
+			data = DataReader(args[position])
+			args = (data, *args[1:])
+
+		return wrapped(*args, **kwargs)
+
+	return wrapper(wrapped)