+++ tests/test_decorators.py
@@ -1,4 +1,8 @@
-from tbm_utils import cast_to_list
+from tbm_utils import (
+	DataReader,
+	cast_to_list,
+	datareader,
+)
 
 
 @cast_to_list(0)
@@ -14,3 +18,29 @@ def cast_func_2(arg0, arg1):
 def test_cast_to_list():
 	assert cast_func_1('test') == cast_func_1(['test']) == ['test']
 	assert cast_func_2('test', 'test') == cast_func_2('test', ['test']) == ('test', ['test'])
+
+
+@datareader
+def func(data):
+	return data
+
+
+class Class:
+	def __init__(self, data):
+		self.data_reader = data
+
+	@datareader
+	@classmethod
+	def build(cls, junk, data):
+		return cls(data)
+
+
+def test_datareader():
+	assert isinstance(func(b'test'), DataReader)
+
+	# c = Class.build(b'junk', b'test')
+	# print(c.datareader)
+	# assert isinstance(c.data_reader, DataReader)
+
+	data_reader = DataReader(b'test')
+	assert func(data_reader) == data_reader